Refactor to include support for ironic + libvirt
This patch centers around two major changes: a rewrite of the way
configuration is handled, and modifications to the virtual BMC object
to allow for the creation of vBMCs with differing functionalities.
In its current state, it is functional but missing a few things
(mentioned below). I wrote this while working for Red Hat as an intern,
and my time ran just short before I could finish all the things I
intended to do. And since I'm leaving this in an unfinished state, I
want to explain as much as I can for whoever may look at this next.

The configuration is now handled by oslo.config, which loads config
options from files and CLI arguments for use by the vbmc client &
server. Since this program defines vBMCs with config files, and since
these configurations become more complex if each one can be of a
different type requiring different option values, a custom oslo.config
ConfigOpts interface was also created to handle these needs. It is
responsible for loading config files, writing them, and providing the
correct arguments to the "vbmc add" subcommand.

I decided to use oslo.config for this because it integrates nicely with
other OpenStack components. In particular, I intended to use
keystoneauth1 to assist with authentication and handling credentials,
but unfortunately, I ran out of time. (see
https://docs.openstack.org/keystoneauth/latest/api/keystoneauth1.loading.conf.html)

Tests for all of these new additions have been written/rewritten,
except for the manager tests, which I also could not finish in time. I
also intended to make the libvirt/ironic-specific features work more
like plugins/extensions that can be enabled or disabled-- primarily to
allow for Ironic vBMCs to be created without needing to install libvirt
dependencies and vice versa.

There's more changes I haven't touched on, I'm sure-- there's a lot
here --so if you have questions, let me know here on GitHub or at
sam11120a@gmail.com. :).